requirements.txt
fastapi~=0.115.2
pytest~=6.2.2
requests~=2.32.0
asyncpg~=0.30.0
pydantic~=2.9.2
python-dotenv~=1.0.1
gotrue~=2.9.3
python-decouple~=3.8

#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via starlette
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via
    #   email-validator
    #   pymongo
email-validator==2.1.0.post1
    # via pydantic
fastapi==0.110.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.7
    # via
    #   anyio
    #   email-validator
motor==3.3.1
    # via -r requirements.in
pydantic==2.6.3
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.16.3
    # via pydantic
pymongo==4.6.3
    # via motor
sniffio==1.3.0
    # via anyio
starlette==0.36.3
    # via fastapi
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.28.0
    # via -r requirements.in